token more fields and broadcast Storedblock

+++ bigtangle-server/src/main/java/net/bigtangle/server/service/schedule/BlockBatchService.java
@@ -5,6 +5,7 @@
 package net.bigtangle.server.service.schedule;
 
 import java.util.List;
+import java.util.concurrent.locks.ReentrantLock;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -22,6 +23,7 @@ import net.bigtangle.server.config.ServerConfiguration;
 import net.bigtangle.server.service.BlockService;
 import net.bigtangle.server.service.TransactionService;
 import net.bigtangle.store.FullPrunedBlockStore;
+import net.bigtangle.utils.Threading;
 
 @Component
 @EnableAsync
@@ -46,43 +48,51 @@ public class BlockBatchService {
 
     @Autowired
     ServerConfiguration serverConfiguration;
-    
+
     @Scheduled(fixedRate = 10000)
 
     public void batch() {
-        if (scheduleConfiguration.isBlockBatchService_active()&& serverConfiguration.checkService()) {
-            batchDo();
+        if (scheduleConfiguration.isBlockBatchService_active() && serverConfiguration.checkService()) {
+            startSingleProcess();
         }
 
     }
 
-    public void batchDo() {
+    protected final ReentrantLock lock = Threading.lock("BlockBatchService");
 
-            logger.info("BlockBatchService start");
-            try {
-                List<BatchBlock> batchBlocks = this.store.getBatchBlockList();
-                if (batchBlocks.isEmpty()) {
-                    return;
-                }
-                Block block = transactionService.askTransactionBlock();
-                for (BatchBlock batchBlock : batchBlocks) {
-                    byte[] payloadBytes = batchBlock.getBlock();
-                    Block putBlock = this.networkParameters.getDefaultSerializer().makeBlock(payloadBytes);
-                    for (Transaction transaction : putBlock.getTransactions()) {
-                        block.addTransaction(transaction);
-                    }
-                }
-                if (block.getTransactions().size() == 0) {
-                    return;
-                }
-                block.solve();
-                blockService.saveBlock(block);
-                for (BatchBlock batchBlock : batchBlocks) {
-                    this.store.deleteBatchBlock(batchBlock.getHash());
-                }
-            } catch (Exception e) {
-                logger.info("BlockBatchService error", e);
-            }
+    public void startSingleProcess() {
+        if (!lock.tryLock()) {
+            logger.debug(this.getClass().getName() + "  Update already running. Returning...");
+            return;
         }
 
+        logger.info("BlockBatchService start");
+        try {
+            List<BatchBlock> batchBlocks = this.store.getBatchBlockList();
+            if (batchBlocks.isEmpty()) {
+                return;
+            }
+            Block block = transactionService.askTransactionBlock();
+            for (BatchBlock batchBlock : batchBlocks) {
+                byte[] payloadBytes = batchBlock.getBlock();
+                Block putBlock = this.networkParameters.getDefaultSerializer().makeBlock(payloadBytes);
+                for (Transaction transaction : putBlock.getTransactions()) {
+                    block.addTransaction(transaction);
+                }
+            }
+            if (block.getTransactions().size() == 0) {
+                return;
+            }
+            block.solve();
+            blockService.saveBlock(block);
+            for (BatchBlock batchBlock : batchBlocks) {
+                this.store.deleteBatchBlock(batchBlock.getHash());
+            }
+        } catch (Exception e) {
+            logger.info("BlockBatchService error", e);
+        } finally {
+            lock.unlock();
+        }
+
+    }
 }
